Refactored the codebase and decomposed it further.
